make save

+++ srcs/history/deload_history.c
@@ -18,14 +18,17 @@ void	ft_deload_hist_in_file(t_shell *sh)
 	char	*path;
 
 	hist = sh->hist;
-	while (hist->next)
-		hist = hist->next;
-	if (!(path = get_tenvv_val(sh->env, "HISTFILE")))
-		return ;
-	unlink(path);
-	while (hist)
+	if (hist)
 	{
+		while (hist->next)
+			hist = hist->next;
+		if (!(path = get_tenvv_val(sh->env, "HISTFILE")))
+			return ;
+		unlink(path);
+		while (hist)
+		{
 			ft_write_in_file(path, hist->s);
-		hist = hist->prev;
+			hist = hist->prev;
+		}
 	}
 }
